+++ src/vertagus/cli/commands/show_alias.py
@@ -6,13 +6,15 @@ from vertagus.configuration import types as cfgtypes
 from vertagus import factory
 from vertagus.cli import utils as cli_utils
 from vertagus.aliases import loader as alias_loader
+from vertagus.cli.options import configless_options
 
 
 @click.command("show-alias")
 @click.argument("alias_name", required=True)
 @click.option("--config", "-c", default=None, help="Path to the configuration file")
-def show_alias_cmd(alias_name, config):
-    master_config = cli_utils.load_config(config, suppress_logging=True)
+@configless_options
+def show_alias_cmd(alias_name, config, **cli_opts):
+    master_config = cli_utils.resolve_config(config, cli_opts, suppress_logging=True)
     project = factory.create_project(cfgtypes.ProjectData.from_project_config(master_config["project"]))
     version = project.get_version()
     if not version: